% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculateFC}
\alias{calculateFC}
\title{Calculate the fold-change}
\usage{
calculateFC(model)
}
\arguments{
\item{model}{nplr model object}
}
\value{
Numeric, fold-change
}
\description{
Calculate the fold-change
}
