% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitiyMetrics.R
\name{calculateMSWD}
\alias{calculateMSWD}
\title{Calculate reduced Chi-squared statistic / mean squared weighted deviation (MSWD)}
\usage{
calculateMSWD(res, internal = TRUE, nConc = 2)
}
\arguments{
\item{res}{Object of class MALDIassay}
}
\value{
Numeric vector of MSWD
}
\description{
Calculate reduced Chi-squared statistic / mean squared weighted deviation (MSWD)
}
\details{
The reduced chi-square statistic is used extensively in goodness of fit testing.
It is also known as mean squared weighted deviation (\strong{MSWD}) in isotopic dating
and variance of unit weight in the context of weighted least squares.

\deqn{\chi^2_\nu =\frac{\chi^2}{\nu}}

\deqn{\chi^2=\sum_{N}{\frac{1}{\sigma^2_i}*[y_i - f(x_i)]^2}}

\deqn{\nu = n - m}
}
