% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRecalibration.R
\name{checkRecalibration}
\alias{checkRecalibration}
\title{Check the recalibration of spectra from a MALDIassay object}
\usage{
checkRecalibration(object, idx)
}
\arguments{
\item{object}{Object of class MALDIassay}

\item{idx}{Numeric, index of spectrum to plot}
}
\value{
ggplot object
}
\description{
Dashed gray lines indicate the mz used for re-calibration ± the tolerance.
Red dashed line indicate the mz used for re-calibration and solid lines indicate peaks.
The spectrum will show the peak used for re-calibration ± 10x the tolerance.
}
\examples{
# see example for `fitCurve()` to see how this data was generated
data(Blank2022res)
checkRecalibration(Blank2022res, idx = 1:8)       
}
