% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdMassSpectrum.R
\name{colSdColMeans}
\alias{colSdColMeans}
\title{Column wise standard deviation}
\usage{
colSdColMeans(x, na.rm = TRUE)
}
\arguments{
\item{x}{matrix}

\item{na.rm}{logical}
}
\value{
sd
}
\description{
also a function of sgibb
see https://stackoverflow.com/questions/17549762/is-there-such-colsd-in-r
}
