% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimates.R
\name{.absToProp}
\alias{.absToProp}
\title{Convert numeric vector into proportions}
\usage{
.absToProp(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector in proportions (0 - 1)
}
\description{
Convert numeric vector into proportions
}
