% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitiyMetrics.R
\name{.calculateSSMD}
\alias{.calculateSSMD}
\title{internal calculation of ssmd}
\usage{
.calculateSSMD(pos, neg)
}
\arguments{
\item{pos}{numeric vector, positive sample}

\item{neg}{numeric vector, negative sample}
}
\value{
ssmd
}
\description{
internal calculation of ssmd
}
