% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectPeaks.R
\name{.detectPeaks}
\alias{.detectPeaks}
\title{Detect peaks}
\usage{
.detectPeaks(spec, SNR, method = "SuperSmoother", halfWindowSize = 3)
}
\arguments{
\item{spec}{MALDIquant::MassSpectrum or list thereof}

\item{SNR}{Numeric, signal to noise value}

\item{method}{Character, method see MALDIquant::detectPeaks}

\item{halfWindowSize}{Numeric, defines width of window for peak detection. See \code{MALDIquant::detectPeaks()}.}
}
\value{
List of MALDIquant::MassPeaks with the same names as \code{spec}
}
\description{
Detect peaks
}
\details{
Just a wrapper around MALDIquant::detectPeaks to ensure that the returned peak list is named.
}
