% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimates.R
\name{.getEstimates}
\alias{.getEstimates}
\title{Get concentration leading to effect size}
\usage{
.getEstimates(model, target)
}
\arguments{
\item{model}{\code{nplr::nplr} object}

\item{target}{Numeric, typically 0.5 for the EC50 or IC50}
}
\value{
Numeric, pICtarget/pECtraget value
}
\description{
Get concentration leading to effect size
}
