% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimates.R
\name{.getFitValue}
\alias{.getFitValue}
\title{Get value of curve for a give target effect size (e.g. 50\%)}
\usage{
.getFitValue(model, target)
}
\arguments{
\item{model}{\code{nplr::nplr} object}

\item{target}{Numeric, typically 0.5 for the EC50 or IC50}
}
\value{
Numeric, value on the fit curve.
Used to provide the correct target for \code{nplr::getEstimates()}
}
\description{
Get value of curve for a give target effect size (e.g. 50\%)
}
