% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMzShift.R
\name{.getMzShift}
\alias{.getMzShift}
\title{Get mass shift for target mz - internal}
\usage{
.getMzShift(peaksdf, tol, targetMz, tolppm = TRUE, allowNoMatch = TRUE)
}
\arguments{
\item{peaksdf}{data.frame with peaks information as generated by peaks2df()}

\item{tol}{Numeric, tolerance around targetMz}

\item{targetMz}{Numeric, target mass}

\item{tolppm}{Logical, is the tolerance provided in ppm (TRUE) or Dalton (FALSE)}

\item{allowNoMatch}{Logical, stop if targetMz is not found in single spectrum?
If TRUE spectra without targetMz match will be excluded.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        List with two entries:
                                  MzShift The mass shift for each spectrum
                                  specIdx The index of the spectra with a match for targetMz
}\if{html}{\out{</div>}}
}
\description{
Get mass shift for target mz - internal
}
