% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{.preprocess}
\alias{.preprocess}
\title{Internal function to perform preprocessing of spectra}
\usage{
.preprocess(
  peaks_single,
  spec,
  SinglePointRecal,
  normMz,
  normTol,
  normMeth,
  alignTol,
  allowNoMatches,
  verbose = TRUE
)
}
\arguments{
\item{peaks_single}{List of MALDIquant::MassPeaks objects.}

\item{spec}{List of MALDIquant::MassSpectrum objects.}

\item{SinglePointRecal}{Logical, perform single point re-calibration}

\item{normMz}{Numeric, m/z used for normalization and re-calibration.}

\item{normTol}{Numeric, tolerance around \code{normMz} in Da.}

\item{normMeth}{Character, normalization method. Options are: "TIC", "median", "PQN" and "mz".}

\item{alignTol}{Numeric, tolerance for alignment in Da.}

\item{allowNoMatches}{Logical, allow no matches for normalization using "mz" method and/or re-calibration.}

\item{verbose}{Logical, print logs to the console.}
}
\description{
Internal function to perform preprocessing of spectra
}
