% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repairMetaData.R
\name{.repairMetaData}
\alias{.repairMetaData}
\title{Repair problematic meta data entries of spectra}
\usage{
.repairMetaData(spec)
}
\arguments{
\item{spec}{list of MALDIquant::MassSpectrum}
}
\value{
list of MALDIquant::MassSpectrum with NA entries in meta data substituted with empty strings (unproblematic).
}
\description{
Repair problematic meta data entries of spectra
}
