% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{.scalingFactor}
\alias{.scalingFactor}
\title{.scalingFactor re implemented from sgibb/MALDIquant}
\usage{
.scalingFactor(object, method = c("TIC", "median"), range)
}
\arguments{
\item{object}{AbstractMassObject}

\item{method}{character}

\item{range}{double}
}
\value{
double
}
\description{
.scalingFactor re implemented from sgibb/MALDIquant
}
