% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getAppliedMzShift}
\alias{getAppliedMzShift}
\title{Extract applied mz-shift}
\usage{
getAppliedMzShift(object)
}
\arguments{
\item{object}{Object of class MALDIassay}
}
\value{
Numeric vector of mz-shits applied to spectra
}
\description{
Extract applied mz-shift
}
\examples{
# see example for `fitCurve()` to see how this data was generated
data(Blank2022res)
head(getAppliedMzShift(Blank2022res))
}
