% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getAppliedNormFactors}
\alias{getAppliedNormFactors}
\title{Extract applied normalization factors}
\usage{
getAppliedNormFactors(object)
}
\arguments{
\item{object}{Object of class MALDIassay}
}
\value{
Numeric vector of normalization factors applied to spectra
}
\description{
Extract applied normalization factors
}
\examples{
# see example for `fitCurve()` to see how this data was generated
data(Blank2022res)
head(getAppliedNormFactors(Blank2022res))
}
