% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getAvgPeaks}
\alias{getAvgPeaks}
\title{Extract peaks of average spectra}
\usage{
getAvgPeaks(object)
}
\arguments{
\item{object}{Object of class MALDIassay}
}
\value{
List of MALDIquantMassPeaks
}
\description{
Extract peaks of average spectra
}
\examples{
# see example for `fitCurve()` to see how this data was generated
data(Blank2022res)
getAvgPeaks(Blank2022res)[[1]]
}
