% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getConc}
\alias{getConc}
\title{Extract the concentrations used in a MALDIassay}
\usage{
getConc(object)
}
\arguments{
\item{object}{Object of class MALDIassay}
}
\value{
Numeric vector, concentrations used in a MALDIassay
}
\description{
Extract the concentrations used in a MALDIassay
}
\examples{
# see example for `fitCurve()` to see how this data was generated
data(Blank2022res)
head(getConc(Blank2022res))
}
