% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getDirectory}
\alias{getDirectory}
\title{Extract directory path}
\usage{
getDirectory(object)
}
\arguments{
\item{object}{Object of class MALDIassay}
}
\value{
List, containing the data used to do the fits as well as the nlpr curve fit .
}
\description{
Extract directory path
}
\examples{
# see example for `fitCurve()` to see how this data was generated
data(Blank2022res)
getDirectory(Blank2022res)
}
