% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getFittingParameters}
\alias{getFittingParameters}
\title{Get fitting parameters}
\usage{
getFittingParameters(object, summarise = FALSE)
}
\arguments{
\item{object}{Object of class MALDIassay}

\item{summarise}{Logical, remove everything other then npar and mz from result.}
}
\value{
tibble of fitting parameters for each fitted m/z-value
}
\description{
Get fitting parameters
}
\examples{
# see example for `fitCurve()` to see how this data was generated
data(Blank2022res)
head(getFittingParameters(Blank2022res, summarise = FALSE))
}
