% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getMzFromMzIdx}
\alias{getMzFromMzIdx}
\title{Get the mz value associated with a mzIdx}
\usage{
getMzFromMzIdx(object, mzIdx)
}
\arguments{
\item{object}{Object of class MALDIassay}

\item{mzIdx}{numeric, index of mass of interest (see \code{getPeakStatistics()})}
}
\value{
numeric, mz value
}
\description{
Get the mz value associated with a mzIdx
}
\examples{
# see example for `fitCurve()` to see how this data was generated
data(Blank2022res)
getMzFromMzIdx(Blank2022res, mzIdx = 2)
}
