% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getNormMethod}
\alias{getNormMethod}
\title{Extract normalization method}
\usage{
getNormMethod(object)
}
\arguments{
\item{object}{Object of class MALDIassay}
}
\value{
Character, normalization method used.
}
\description{
Extract normalization method
}
\examples{
# see example for `fitCurve()` to see how this data was generated
data(Blank2022res)
getNormMethod(Blank2022res)
}
