% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getNormMz}
\alias{getNormMz}
\title{Extract m/z used for normalization}
\usage{
getNormMz(object)
}
\arguments{
\item{object}{Object of class MALDIassay}
}
\value{
Numeric, m/z used for normalization
}
\description{
Extract m/z used for normalization
}
\examples{
# see example for `fitCurve()` to see how this data was generated
data(Blank2022res)
getNormMz(Blank2022res) 
}
