% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getPeakStatistics}
\alias{getPeakStatistics}
\title{Extract peak statistics}
\usage{
getPeakStatistics(object, summarise = FALSE)
}
\arguments{
\item{object}{Object of class MALDIassay}

\item{summarise}{Logical, return summarized results (one result per mz and not per mz and spectra)}
}
\value{
A tibble with peak statistics (R², fold-change, CV\%, etc.)
}
\description{
Extract peak statistics
}
\examples{
# see example for `fitCurve()` to see how this data was generated
data(Blank2022res)
head(getPeakStatistics(Blank2022res, summarise = TRUE))
}
