% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getSinglePeaks}
\alias{getSinglePeaks}
\title{Extract peaks of single spectra spectra (before average calculation)}
\usage{
getSinglePeaks(object)
}
\arguments{
\item{object}{Object of class MALDIassay}
}
\value{
List of MALDIquantMassPeaks
}
\description{
Extract peaks of single spectra spectra (before average calculation)
}
\examples{
# see example for `fitCurve()` to see how this data was generated
data(Blank2022res)
getSinglePeaks(Blank2022res)[[1]]
}
