% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getSingleSpecIntensity}
\alias{getSingleSpecIntensity}
\title{Extract the intensities of single spectra for a given mzIdx}
\usage{
getSingleSpecIntensity(object, mz_idx)
}
\arguments{
\item{object}{Object of class MALDIassay}

\item{mz_idx}{Integer, index of mz}
}
\value{
Numeric vector, intensities of mzIdx
}
\description{
Extract the intensities of single spectra for a given mzIdx
}
\examples{

# see example for `fitCurve()` to see how this data was generated
data(Blank2022res)
head(getSingleSpecIntensity(Blank2022res, 2))
}
