% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getSNR}
\alias{getSNR}
\title{Extract SNR used for peak detection}
\usage{
getSNR(object)
}
\arguments{
\item{object}{Object of class MALDIassay}
}
\value{
Numeric, SNR used for peak detection
}
\description{
Extract SNR used for peak detection
}
\examples{
# see example for `fitCurve()` to see how this data was generated
data(Blank2022res)
getSNR(Blank2022res)
}
