% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getVarFilterMethod}
\alias{getVarFilterMethod}
\title{Extract variance filtering method}
\usage{
getVarFilterMethod(object)
}
\arguments{
\item{object}{Object of class MALDIassay}
}
\value{
Character of variance filtering method used
}
\description{
Extract variance filtering method
}
\examples{
# see example for `fitCurve()` to see how this data was generated
data(Blank2022res)
getVarFilterMethod(Blank2022res)
}
