% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{isMALDIassay}
\alias{isMALDIassay}
\title{Check if object if of class MALDIassay}
\usage{
isMALDIassay(object)
}
\arguments{
\item{object}{object to text}
}
\value{
logical, TRUE if object is of class MALDIassay
}
\description{
Check if object if of class MALDIassay
}
\examples{
x <- 1
# FALSE
isMALDIassay(x)
# TRUE
isMALDIassay(Blank2022res)
}
