% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MALDIassay_Class.R
\docType{class}
\name{MALDIassay-class}
\alias{MALDIassay-class}
\title{Class MALDIassay}
\arguments{
\item{object}{MALDIassay.}
}
\description{
A class for holding MALDI assay related information.
}
