% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks2df.R
\name{peaks2df}
\alias{peaks2df}
\title{Convert a list of peaks to a data.frame}
\usage{
peaks2df(peaks)
}
\arguments{
\item{peaks}{(list of) MALDIquant::MassPeaks}
}
\value{
Data.frame with peak data
}
\description{
Convert a list of peaks to a data.frame
}
\examples{
data(Blank2022peaks)

peakdf <- peaks2df(Blank2022peaks[1:2])
head(peakdf)

}
