% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plotOverview}
\alias{plotOverview}
\title{Generate an overview of a fitted curves from a MALDIassay object}
\usage{
plotOverview(object)
}
\arguments{
\item{object}{object of class MALDIassay}
}
\value{
arranged plots
}
\description{
Generate an overview of a fitted curves from a MALDIassay object
}
