% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdMassSpectrum.R
\name{sdMassSpectraFun}
\alias{sdMassSpectraFun}
\title{Still fork from sgibb MALDIquant}
\usage{
sdMassSpectraFun(l, mergeMetaData = TRUE)
}
\arguments{
\item{l}{see MALDIquant}

\item{mergeMetaData}{see MALDIquant}

\item{fun}{see MALDIquant}
}
\value{
see MALDIquant
}
\description{
Still fork from sgibb MALDIquant
}
