% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spec}
\alias{spec}
\title{MALDI mass spectra from SAHA treated cells}
\format{
List of MALDIquant::MassSpectrum
}
\usage{
spec
}
\description{
Cells were treated with the following concentrations of SAHA:
0, 0.04, 0.12, 0.37, 1.11, 3.33, 10, 30 µM
For each concentration there are 4 replicates ("measurement replicates").
}
\keyword{datasets}
