% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabilizeVariance.R
\name{stabilizeVariance}
\alias{stabilizeVariance}
\title{Stabilize variance in spectra}
\usage{
stabilizeVariance(
  spec,
  method = c("sqrt", "log", "log2", "log10"),
  correctBaseline = c(NA, "SNIP", "TopHat", "ConvexHull", "median")
)
}
\arguments{
\item{spec}{List of MALDIquant::MassSpectrum}

\item{method}{Method for stabilization}

\item{correctBaseline}{Baseline correction method}
}
\value{
stabilized spectra
}
\description{
Stabilize variance in spectra
}
