% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{stopIfNotIsMALDIassay}
\alias{stopIfNotIsMALDIassay}
\title{Stop and throw an error if object is not of class MALDIassay}
\usage{
stopIfNotIsMALDIassay(object)
}
\arguments{
\item{object}{object}
}
\description{
Stop and throw an error if object is not of class MALDIassay
}
