% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{transformConc2Log}
\alias{transformConc2Log}
\title{Convert concentration to log10 and replace zero's}
\usage{
transformConc2Log(conc)
}
\arguments{
\item{conc}{numeric, concentrations.}
}
\value{
numeric, log10 transformed concentrations
}
\description{
Convert concentration to log10 and replace zero's
}
\examples{
transformConc2Log(c(0.1, 0.01,0.001))
}
