% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCurveFit.R
\name{calculateCurveFit}
\alias{calculateCurveFit}
\title{Calculate the fit for a dose-response curve}
\usage{
calculateCurveFit(intmat, idx, npars = 4, ...)
}
\arguments{
\item{intmat}{Intensity matrix as generated by
\code{MALDIquant::intensityMatrix()}
with rownames as the respectiv concentrations of the spectra.}

\item{idx}{Numeric vector of the mz indicies to perform the fit.}

\item{npars}{Numeric or "all", see \code{nplr::nplr()}.}

\item{...}{Additional arguments passed to \code{nplr::nplr()}}
}
\value{

}
\description{
Calculate the fit for a dose-response curve
}
