% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculateFC}
\alias{calculateFC}
\title{Calculate the fold-change}
\usage{
calculateFC(df, absolute = FALSE, n_conc = 2)
}
\arguments{
\item{df}{tibble with columns "conc" and "value".}

\item{absolute}{logical, return the absolute FC.
If FALSE (default) the FC is calculated by dividing the mean intensity of the highest concentrations by the mean intensity of the lowest concentrations.}

\item{n_conc}{numeric, the FC is calculated on the top and bottom \code{n_conc} concentrations (default 2).}
}
\value{
Numeric, fold-change
}
\description{
Calculate the fold-change
}
