% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePeakstatistics.R
\name{calculatePeakStatistics}
\alias{calculatePeakStatistics}
\title{Calculate peak statistics}
\usage{
calculatePeakStatistics(curveFits, intensityMatrix)
}
\arguments{
\item{curveFits}{list of curve fits as returned by ```MALDIcellassay::calculateCurveFit()```}

\item{intensityMatrix}{matrix of intensities with rownames as concentrations of the respective spectra}
}
\value{
A tibble with peak statistics
}
\description{
Calculate peak statistics
}
