% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRecalibration.R
\name{checkRecalibration}
\alias{checkRecalibration}
\title{Check the recalibration of spectra from a MALDIassay object}
\usage{
checkRecalibration(object, idx)
}
\arguments{
\item{object}{Object of class MALDIassay}

\item{idx}{Numeric, index of spectrum to plot}
}
\value{
ggplot object
}
\description{
Dashed grey lines indicate the mz used for recalibration ± the tolerance.
Red dashed line indicate the mz used for recalibration and solid dashed lines indicate peaks.
}
