% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateSpectra.R
\name{.aggregateSpectra}
\alias{.aggregateSpectra}
\title{Preprocessing function to average spectra, detect and bin peaks before turning them into an intensity matrix}
\usage{
.aggregateSpectra(
  spec,
  averageMethod,
  SNR,
  monoisotopicFilter,
  binTol,
  normMz,
  normTol,
  halfWindowSize,
  peakMethod = "SuperSmoother",
  verbose = TRUE
)
}
\arguments{
\item{spec}{List of MALDIquant::MassSpectrum}

\item{averageMethod}{Character, method for aggregation: "mean", "median" or "sum"}

\item{SNR}{Numeric, Signal noise value for peak detection}

\item{monoisotopicFilter}{Logical, filter monoisotopic peaks}

\item{binTol}{Numeric, tolerance for binning}

\item{normMz}{Numeric, m/z value used for normalization/re-calibration}

\item{normTol}{Numeric, tolerance around \code{normMz} in Da.}

\item{halfWindowSize}{Numeric, halfWindowSize for peak detection.}

\item{peakMethod}{Character, method for peak detection. Either "SuperSmoother" or "MAD".}

\item{verbose}{Logical, print logs to the console.}
}
\value{
List of lists with intensity matrix, average spectra and average peaks
}
\description{
Preprocessing function to average spectra, detect and bin peaks before turning them into an intensity matrix
}
