% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractIntensity.R
\name{extractIntensity}
\alias{extractIntensity}
\title{Extract intensity using peaks as template}
\usage{
extractIntensity(peaks, spec, tol = 0.1)
}
\arguments{
\item{peaks}{MALDIquant::MassPeaks list}

\item{spec}{MALDIquant::MassSpectrum list}

\item{tol}{numeric, tolerance in Da}
}
\value{
MALDIquant::MassPeaks list with extracted intensities from spec at m/z of peaks = pseudo peaks.
Useful in combination with sdMassSpectrum to get standard deviation of peaks as intensity matrix.
}
\description{
Extract intensity using peaks as template
}
