% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVariance.R
\name{filterVariance}
\alias{filterVariance}
\title{Filter for high variance signals}
\usage{
filterVariance(vars, method = c("mean", "median", "q25", "q75", "none"))
}
\arguments{
\item{vars}{Numeric vector, veriances of signals}

\item{method}{Character, filtering method. One of "mean" (default), "median", "q25", "q75" (25 and 75\% quantile) or "none".}
}
\value{
Indicies of spectra with a high variance
}
\description{
Filter for high variance signals
}
