% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCurve.R
\name{fitCurve}
\alias{fitCurve}
\title{Fit dose-response curves}
\usage{
fitCurve(
  spec,
  dir,
  conc = NA,
  normMz = 760.585,
  normTol = 0.1,
  alignTol = 0.05,
  binTol = 0.25,
  SNR = 3,
  allowNoMatches = TRUE,
  normMeth = c("mz", "TIC"),
  saveIntensityMatrix = TRUE,
  SinglePointRecal = TRUE
)
}
\arguments{
\item{spec}{List of MALDIquant::MassSpectrum}

\item{dir}{Character, directory for output}

\item{conc}{Numeric vector, concentration for each spectrum. Length has to be the same as length of spec}

\item{normMz}{Numeric, mz used for normalization AND for single point recalibration.}

\item{normTol}{Numeric, tolerance in Dalton to match normMz}

\item{alignTol}{Numeric, tolerance for spectral alignment in Dalton.}

\item{binTol}{Numeric, tolerance for binning of peaks. D}

\item{SNR}{Numeric, signal to noise ratio for peak detection.}

\item{allowNoMatches}{Logical, if normMz can not be found in a spectrum, proceed and exclude spectrum or stop}

\item{normMeth}{Character, normalization method. Can either be "TIC" or "mz". If "mz" then the normMz is used.}

\item{saveIntensityMatrix}{Logical, save the intensity matrix as xlsx to the disk.}

\item{SinglePointRecal}{Logical, perform single point recalibration to normMz}
}
\value{

}
\description{
Fit dose-response curves
}
