% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCurve.R
\name{fitCurve}
\alias{fitCurve}
\title{Fit dose-response curves}
\usage{
fitCurve(
  spec,
  dir,
  conc = NA,
  varFilterMethod = c("mean", "median", "q25", "q75", "none"),
  normMz = 760.585,
  normTol = 0.1,
  alignTol = 0.01,
  binTol = 0.25,
  SNR = 3,
  allowNoMatches = TRUE,
  normMeth = c("mz", "TIC", "PQN", "median", "none"),
  saveIntensityMatrix = TRUE,
  SinglePointRecal = TRUE,
  fc_thresh = 1,
  markValue = NA,
  plot = TRUE
)
}
\arguments{
\item{spec}{List of MALDIquant::MassSpectrum}

\item{dir}{Character, directory for output}

\item{conc}{Numeric vector, concentration for each spectrum. Length has to be the same as length of spec}

\item{varFilterMethod}{Character, function applied for high variance filtering. One of the following options `mean` (default), `median`, `q25`, `q75` or `none` (no filtering).}

\item{normMz}{Numeric, mz used for normalization AND for single point recalibration.}

\item{normTol}{Numeric, tolerance in Dalton to match normMz}

\item{alignTol}{Numeric, tolerance for spectral alignment in Dalton.}

\item{binTol}{Numeric, tolerance for binning of peaks.}

\item{SNR}{Numeric, signal to noise ratio for peak detection.}

\item{allowNoMatches}{Logical, if normMz can not be found in a spectrum, proceed and exclude spectrum or stop}

\item{normMeth}{Character, normalization method. Can either be "TIC", "PQM", "median" or "mz". If "mz" then the normMz is used. If none no normalization is done.}

\item{saveIntensityMatrix}{Logical, save the intensity matrix as xlsx to the disk.}

\item{SinglePointRecal}{Logical, perform single point recalibration to normMz}

\item{fc_thresh}{Numeric, threshold for fold change above which curves are plotted. The fold-chage is calculated as max/min for a given m/z.}

\item{markValue}{Numeric, value to mark in the resulting plot. Set to NA if no value needs to be marked.}

\item{plot}{Logical, should the curves be plotted and written to dir?}
}
\value{

}
\description{
Fit dose-response curves
}
