% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getAvgSpectra}
\alias{getAvgSpectra}
\title{Extract average spectra}
\usage{
getAvgSpectra(object)
}
\arguments{
\item{object}{Object of class MALDIassay}
}
\value{
List of MALDIquantMassSpectrum
}
\description{
Extract average spectra
}
