% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getBinTol}
\alias{getBinTol}
\title{Get binning tolerance}
\usage{
getBinTol(object)
}
\arguments{
\item{object}{Object of class MALDIassay}
}
\value{
Numeric, tolerance used for binning in Dalton.
}
\description{
Get binning tolerance
}
