% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConc.R
\name{getConc}
\alias{getConc}
\title{Extract the concentrations used in a MALDIassay}
\usage{
getConc(object)
}
\arguments{
\item{object}{Object of class MALDIassay}
}
\value{
Concentrations used in a MALDIassay
}
\description{
Extract the concentrations used in a MALDIassay
}
