% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getCurveFits}
\alias{getCurveFits}
\title{Extract curve fits}
\usage{
getCurveFits(object)
}
\arguments{
\item{object}{Object of class MALDIassay}
}
\value{
List, containing the data used to do the fits as well as the nlpr curve fit .
}
\description{
Extract curve fits
}
