% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getFittingParameters}
\alias{getFittingParameters}
\title{Get fitting parameters}
\usage{
getFittingParameters(object, summarise = FALSE)
}
\arguments{
\item{object}{Object of class MALDIassay}
}
\value{
tibble of fitting parameters for each fitted m/z-value
}
\description{
Get fitting parameters
}
