% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getIntensityMatrix}
\alias{getIntensityMatrix}
\title{Get the intensity matrix of single spectra for all fitted curves}
\usage{
getIntensityMatrix(object)
}
\arguments{
\item{object}{Object of class MALDIassay}
}
\value{
A matrix with columns as *m/z* values and rows as concentrations/spectra
}
\description{
Get the intensity matrix of single spectra for all fitted curves
}
\details{
Note that the returned matrix only contains *m/z* values that were actually fitted.
If a variance filtering step was applied this will not include **all** *m/z* values.
If you wish to get a matrix of **all** *m/z* values use ```MALDIquant::intensityMatrix(getSinglePeaks(object))```.
}
