% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getNormMzTol}
\alias{getNormMzTol}
\title{Extract tolerance used for normalization}
\usage{
getNormMzTol(object)
}
\arguments{
\item{object}{Object of class MALDIassay}
}
\value{
Numeric, tolerance used for normalization
}
\description{
Extract tolerance used for normalization
}
