% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getPeakStatistics}
\alias{getPeakStatistics}
\title{Extract peak statistics}
\usage{
getPeakStatistics(object)
}
\arguments{
\item{object}{Object of class MALDIassay}
}
\value{
A tibble with peak statistics (R², fold-change, CV%, etc.)
}
\description{
Extract peak statistics
}
