% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRecalibration.R
\name{getRecalibrationError}
\alias{getRecalibrationError}
\title{Calculate remaining calibration error of a MALDIassay object}
\usage{
getRecalibrationError(object)
}
\arguments{
\item{object}{Object of class MALDIassay}
}
\value{
A tibble containing statistics about remaining calibration error
}
\description{
Calculate remaining calibration error of a MALDIassay object
}
