% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getSingleSpecIntensity}
\alias{getSingleSpecIntensity}
\title{Extract the intensities of single spectra for a given mzIdx}
\usage{
getSingleSpecIntensity(object, mzIdx)
}
\arguments{
\item{object}{Object of class MALDIassay}

\item{mzIdx}{Integer, index of mz}
}
\value{
Numeric vector, intensities of mzIdx
}
\description{
Extract the intensities of single spectra for a given mzIdx
}
