% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{isMALDIassay}
\alias{isMALDIassay}
\title{Check if object if of class MALDIassay}
\usage{
isMALDIassay(object)
}
\arguments{
\item{object}{object}
}
\value{
logical
}
\description{
Check if object if of class MALDIassay
}
