% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSpectra.R
\name{loadSpectraMzML}
\alias{loadSpectraMzML}
\title{load mzML spectra}
\usage{
loadSpectraMzML(
  Dir,
  filter = NA,
  nameSpectra = TRUE,
  verbose = TRUE,
  centroided = FALSE
)
}
\arguments{
\item{Dir}{Character, parent directory of spectra.}

\item{filter}{Character vector, filter out spectra which match the given vector.}

\item{nameSpectra}{Logical, if TRUE the spectra in the resulting list will be named according to the dirname.}

\item{verbose}{Logical, print logs to console}

\item{centroided}{Logical, if FALSE (default) a list of MALDIquant::MassSpectrum is returned, otherwise a list of MALDIquant::MassPeaks are returned.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       List of MALDIquant::MassSpectra
}\if{html}{\out{</div>}}
}
\description{
load mzML spectra
}
\examples{
dataDir <- system.file("extdata", package="MALDIcellassay")

loadSpectraMzML(file.path(dataDir, "Koch2024mzML"))
}
