% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize spectra and peaks}
\usage{
normalize(spec, peaks, normMeth)
}
\arguments{
\item{spec}{List of MALDIquant::MassSpectrum}

\item{peaks}{List of MALDIquant::MassPeaks}

\item{normMeth}{Character, normalization method. Options are "TIC", "PQN", "median" and "mz".}
}
\value{
List of lists of normalized MALDIquant::MassSpectrum and normalized MALDIquant::MassPeaks
}
\description{
Normalize spectra and peaks
}
