% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeByFactor.R
\name{normalizeByFactor}
\alias{normalizeByFactor}
\title{Apply normalization factors to spectra}
\usage{
normalizeByFactor(spec, factors)
}
\arguments{
\item{spec}{List of MALDIquant::MassSpectrum or MALDIquant::MassPeaks}

\item{factors}{Numeric vector of normalization factors. See getNormFactors().}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        List of normalized Spectra or Peaks
}\if{html}{\out{</div>}}
}
\description{
Apply normalization factors to spectra
}
