% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCurves.R
\name{plotCurves}
\alias{plotCurves}
\title{generate ggplot objects for each of the curve fits in a MALDIassay object}
\usage{
plotCurves(object, fc_thresh = 1, markValue = NA)
}
\arguments{
\item{object}{object of class MALDIassay}

\item{fc_thresh}{numeric, max/min fold-change above which plots should be generated}

\item{markValue}{numeric, a value to add as a reference to the plots}
}
\value{
list of ggplot objects
}
\description{
generate ggplot objects for each of the curve fits in a MALDIassay object
}
