% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPeak.R
\name{plotPeak}
\alias{plotPeak}
\title{Plot a peak of interest from a MALDIassay object}
\usage{
plotPeak(object, mzIdx, tol = 4)
}
\arguments{
\item{object}{object of class MALDIassay}

\item{mzIdx}{numeric, index of mass of interest (see \code{getPeakStatistics()})}

\item{tol}{numeric, tolerance around peak to plot}
}
\value{
ggplot object
}
\description{
Plot a peak of interest from a MALDIassay object
}
