% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{savePlots}
\alias{savePlots}
\title{Write ggplot objects to disk as png}
\usage{
savePlots(object, dir = NULL)
}
\arguments{
\item{object}{object of class MALDIassay}
}
\description{
Write ggplot objects to disk as png
}
